Move topic page config to the correct location and add tests
