+++ src/app/routes/homePage/getInitialData/page-config.ts
@@ -1,18 +1,6 @@
-import { Services, Environments } from '../../../models/types/global';
+import { Services } from '../../../models/types/global';
 
 type Envs = 'test' | 'live';
-
-export type TopicPagePaths = '/persian/afghanistan';
-
-export const TOPIC_PAGE_CONFIG: {
-  [path in TopicPagePaths]: { [env in Environments]: string };
-} = {
-  '/persian/afghanistan': {
-    local: 'crezq2dg9zwt',
-    test: 'c15er11zq57t',
-    live: 'crezq2dg9zwt',
-  },
-};
 
 export const HOME_PAGE_CONFIG: {
   [service in Services]: { [env in Envs]: string };
